% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_generate_norms.R
\name{generate_norms}
\alias{generate_norms}
\title{Calculate all the norms from a Conceptual properties}
\usage{
generate_norms(orig_data)
}
\arguments{
\item{orig_data}{Data frame with 3 columns: id, concept and properties}
}
\value{
Data frame with all the estimations of norms
}
\description{
Calculate all the norms from a Conceptual properties
}
\examples{
data_test = data.frame(CPN_27)
generate_norms(data_test)
}
