% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLT_app.R
\name{WordListsAnalytics}
\alias{WordListsAnalytics}
\title{PLT App function}
\usage{
WordListsAnalytics()
}
\value{
None (it executes a shiny application).
}
\description{
Main function of package. It executes a shiny application in local session.
The user can load data, generate new descriptive words, apply a new clustering model, and
use simulations to estimate the probability that two persons describe the same words
based on their descriptions.
}
\examples{
if(interactive()){
  WordListsAnalytics()
}
}
